subdirfs: renamed to subdirfs
