+++ subdirfs/subdir_test.go
@@ -1,4 +1,4 @@
-package subdir
+package subdirfs
 
 import (
 	"io/ioutil"
